+++ src/config/modelProviders/qwen.ts
@@ -74,6 +74,7 @@ const Qwen: ModelProviderCard = {
     },
   ],
   checkModel: 'qwen-turbo',
+  disableBrowserRequest: true,
   id: 'qwen',
   modelList: { showModelFetcher: true },
   name: 'Qwen',