+++ src/libs/agent-runtime/AgentRuntime.ts
@@ -145,93 +145,93 @@ class AgentRuntime {
       }
 
       case ModelProvider.ZhiPu: {
-        runtimeModel = await LobeZhipuAI.fromAPIKey(params.zhipu ?? {});
+        runtimeModel = await LobeZhipuAI.fromAPIKey(params.zhipu);
         break;
       }
 
       case ModelProvider.Google: {
-        runtimeModel = new LobeGoogleAI(params.google ?? {});
+        runtimeModel = new LobeGoogleAI(params.google);
         break;
       }
 
       case ModelProvider.Moonshot: {
-        runtimeModel = new LobeMoonshotAI(params.moonshot ?? {});
+        runtimeModel = new LobeMoonshotAI(params.moonshot);
         break;
       }
 
       case ModelProvider.Bedrock: {
-        runtimeModel = new LobeBedrockAI(params.bedrock ?? {});
+        runtimeModel = new LobeBedrockAI(params.bedrock);
         break;
       }
 
       case ModelProvider.Ollama: {
-        runtimeModel = new LobeOllamaAI(params.ollama ?? {});
+        runtimeModel = new LobeOllamaAI(params.ollama);
         break;
       }
 
       case ModelProvider.Perplexity: {
-        runtimeModel = new LobePerplexityAI(params.perplexity ?? {});
+        runtimeModel = new LobePerplexityAI(params.perplexity);
         break;
       }
 
       case ModelProvider.Anthropic: {
-        runtimeModel = new LobeAnthropicAI(params.anthropic ?? {});
+        runtimeModel = new LobeAnthropicAI(params.anthropic);
         break;
       }
 
       case ModelProvider.DeepSeek: {
-        runtimeModel = new LobeDeepSeekAI(params.deepseek ?? {});
+        runtimeModel = new LobeDeepSeekAI(params.deepseek);
         break;
       }
 
       case ModelProvider.Minimax: {
-        runtimeModel = new LobeMinimaxAI(params.minimax ?? {});
+        runtimeModel = new LobeMinimaxAI(params.minimax);
         break;
       }
 
       case ModelProvider.Mistral: {
-        runtimeModel = new LobeMistralAI(params.mistral ?? {});
+        runtimeModel = new LobeMistralAI(params.mistral);
         break;
       }
 
       case ModelProvider.Groq: {
-        runtimeModel = new LobeGroq(params.groq ?? {});
+        runtimeModel = new LobeGroq(params.groq);
         break;
       }
 
       case ModelProvider.OpenRouter: {
-        runtimeModel = new LobeOpenRouterAI(params.openrouter ?? {});
+        runtimeModel = new LobeOpenRouterAI(params.openrouter);
         break;
       }
 
       case ModelProvider.TogetherAI: {
-        runtimeModel = new LobeTogetherAI(params.togetherai ?? {});
+        runtimeModel = new LobeTogetherAI(params.togetherai);
         break;
       }
 
       case ModelProvider.ZeroOne: {
-        runtimeModel = new LobeZeroOneAI(params.zeroone ?? {});
+        runtimeModel = new LobeZeroOneAI(params.zeroone);
         break;
       }
 
       case ModelProvider.Qwen: {
-        runtimeModel = new LobeQwenAI(params.qwen ?? {});
+        runtimeModel = new LobeQwenAI(params.qwen);
         break;
       }
 
       case ModelProvider.Stepfun: {
-        runtimeModel = new LobeStepfunAI(params.stepfun ?? {});
+        runtimeModel = new LobeStepfunAI(params.stepfun);
         break;
       }
 
       case ModelProvider.Baichuan: {
-        runtimeModel = new LobeBaichuanAI(params.baichuan ?? {});
-        break
+        runtimeModel = new LobeBaichuanAI(params.baichuan);
+        break;
       }
 
       case ModelProvider.Taichu: {
-        runtimeModel = new LobeTaichuAI(params.taichu ?? {});
-        break
+        runtimeModel = new LobeTaichuAI(params.taichu);
+        break;
       }
     }